+++ NEWS.md
@@ -2,6 +2,9 @@
   * Bug: .Rbuildignore files are case insensitive!
   * Bug: Parsing .gitlab-ci files
   * Bug: Remove deleted files in pre-commit check
+  * Feature: Raise an error on when `readLines()` doesn't end with a new line
+  * Feature: Check for windows line breaks
+  * Feature: Check file permission
 
 # Version 0.4.0
   * Feature: Add env variable to check_meta